Factor init sequence out of _PG_init into a restartable routine.
The initsequencer routine can pick up where it left off, whether
started by _PG_init, by the internal call handler, or by a GUC hook
if the admin is trying a different config setting.

To do: continue moving bits out of the initializeJavaVM, etc., routines
into additional stages in the sequencer. Eventually: have a stage where
the JVM is running and a Java method can be called to check for and do
the SQL function declarations, SQLJ schema population, etc.