Moved most initialization to initsequencer.
This is where it gets hairy, because it _seemed_ ok to reorder some things,
and some bits have been moved out from under a `PG_TRY`. It successfully
does about what it did before this commit, but the ability to re-create
the VM to try a new classpath has not appeared yet; hotspot crashes on
the second create attempt.