Add pljava.enable to simplify <9.2 installs.
As explained in the added docs....

Also, alter the errhint for saving the settings in the <9.2 case.
Saving the settings with ALTER DATABASE ... SET won't work there, they
have to go in postgresql.conf.

Also masks the exact postgresql.conf path if the user isn't superuser.
I'm fairly sure there's no code path to produce this message if the user
isn't superuser, but a simple check is more readable than a proof.