The Strawberry PG installation is an impostor
The Windows runner seems to have an extra pg_config somewhere on the
path, that reports it was built with MinGW and installed in paths
containing Strawberry that don't really exist.

It has cropped up for others too in other settings:
https://github.com/timescale/timescaledb/issues/555

The pg_config found in $PGBIN refers to a PostgreSQL that was built
with MSVC, and its reported directories really exist. Specify that one
explicitly when running on Windows.

In passing, avoid trying to run the jshell install-and-test with Java 9
in Windows (it will wait forever for someone to type on the console).