Update ci-sourcepg.yml and add ci-packagedpg.yml
Structure the matrix in each the same as in ci-runnerpg.yml
(except that both of these will have a choice of PostgreSQL
version, rather than being stuck with the version supplied
on the runner).

Both of these need to be completed, of course.

There is much duplication among these three workflows (pretty
much everything after installing or building PostgreSQL). They
will probably be easier to get working that way. After most of
the tricky bits have been worked out, it may be possible to
combine them all into one giant uberworkflow with a lot of
conditional steps. That feels like premature optimization for now.