Will this change the result?
The structure of PG_VERSION_NUM changed %d%04d (major, minor)
for PG 10 (well really for 10.1, when it first mattered, see
958fe54). This version conditional wasn't written in the new
style. (Pretend you didn't see it was wrong for the old style
too.)

If this corrects the behavior, it will mean there was something
unreliable about the workaround of getting the timezone name and
passing it to pg_tzset.