The barest start at implementing some FDW API
In a new org.postgresql.pljava.example.fdw package, add two new examples:

DummyFDW is an example FDW implementation: a class that implements
PLJavaBasedFDW (and does what could be fairly described as nothing).
In its SQLAction annotation it creates two SQL-declared functions,
a handler and a validator, both referring to the class by name;
these functions use the 'pljavafdwhandler' language.

FDWHandlerLanguage is a PLJavaBasedLanguage implementation that only
allows two kinds of routine to be declared: an FDW handler and an FDW
validator; for either kind, the AS string is just the name of a class
that implements PLJavaBasedFDW. This, of course, is the pljavafdwhandler
language referred to above. Its existence among the examples is temporary;
ultimately, it should be created in the normal course of PL/Java
installation.

FDWHanderLanguage's implementation of validation is more or less complete.
The handler implementation is more or less entirely vapor.